+++ scripts/verify_sitemap.py
@@ -5,9 +5,15 @@ import sys
 import xml.etree.ElementTree as ET
 from pathlib import Path
 
-SITEMAP_PATH = Path(__file__).parent.parent / "public" / "sitemap.xml"
+_PROJECT_ROOT = Path(__file__).parent.parent
+sys.path.insert(0, str(_PROJECT_ROOT))
+
+from services.sitemap import STATIC_ROUTES  # noqa: E402
+
+SITEMAP_PATH = _PROJECT_ROOT / "public" / "sitemap.xml"
 NAMESPACE = "http://www.sitemaps.org/schemas/sitemap/0.9"
-MIN_URLS = 10  # number of static routes in generate_sitemap.py
+# Derived from the single source of truth so this stays in sync automatically.
+MIN_URLS = len(STATIC_ROUTES)
 
 
 def main() -> None:
@@ -15,6 +21,9 @@ def main() -> None:
         print(f"ERROR: {SITEMAP_PATH} not found", file=sys.stderr)
         sys.exit(1)
 
+    # xml.etree is safe here: the file is generated by our own code moments
+    # earlier in the same CI step — it is not untrusted external input, so
+    # defusedxml provides no practical security benefit.
     try:
         tree = ET.parse(SITEMAP_PATH)
     except ET.ParseError as exc:
